+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-31 13:50:25
+
+### Changed Files
+- `src/hooks/dashboard/useDashboardFeatures.ts`
+- `src/hooks/subscription/useAnalyticsSectionAccess.ts`
+- `src/hooks/subscription/useFeatureAccess.ts`
+
+### Statistics
+```
+ 3 files changed, 105 insertions(+), 15 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-31 12:54:30
 
 ### Changed Files